requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==2.2.3
django-allauth==0.44.0
django-crispy-forms==1.11.2
django-filter==2.4.0
django-forms-bootstrap==3.1.0
django-google-translate==1.1
django-multiselectfield==0.1.12
django-upload-form==0.5.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
fontawesome-free==5.15.3
idna==2.10
oauthlib==3.1.0
Pillow==8.2.0
pycparser==2.20
PyJWT==2.0.1
python-social-auth==0.2.21
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.4
